# openapi/cognigate-api.yaml
openapi: 3.1.0
info:
  title: Cognigate API
  description: |
    # Cognigate AI Governance API

    Cognigate provides AI governance, trust scoring, and policy enforcement for autonomous agents.

    ## Base URLs
    - **Production**: `https://cognigate.dev/v1`
    - **Staging**: `https://staging.cognigate.dev/v1`

    ## Authentication
    All endpoints require Bearer token authentication:
    ```
    Authorization: Bearer <your-api-key>
    ```

    ## Trust Tiers (BASIS Specification)
    | Tier | Score Range | Name | Description |
    |------|-------------|------|-------------|
    | T0 | 0-199 | Sandbox | Isolated, no external access |
    | T1 | 200-349 | Observed | Read-only, monitored |
    | T2 | 350-500 | Provisional | Basic operations, supervised |
    | T3 | 501-649 | Monitored | Standard operations with monitoring |
    | T4 | 650-799 | Standard | External API access, policy-governed |
    | T5 | 800-875 | Trusted | Cross-agent communication |
    | T6 | 876-950 | Certified | Admin tasks, minimal oversight |
    | T7 | 951-1000 | Autonomous | Full autonomy, self-governance |

    ## Related Resources
    - [BASIS Specification](https://basis.vorion.org)
    - [TypeScript SDK](https://npmjs.com/package/@vorionsys/cognigate)
    - [Agent Anchor Trust](https://trust.agentanchorai.com)

  version: 1.0.0
  contact:
    name: Agent Anchor AI
    email: api-support@agentanchorai.com
    url: https://cognigate.dev
  license:
    name: MIT
...
  x-logo:
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /agents/{agentId}:
    get:
      tags: [Agents]
      summary: Get agent
      description: Retrieve agent details
      operationId: getAgent
      parameters:
        - $ref: '#/components/parameters/agentId'
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Agents]
      summary: Update agent
      description: Update agent properties
      operationId: updateAgent
      parameters:
        - $ref: '#/components/parameters/agentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAgentRequest'
      responses:
        '200':
          description: Agent updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Agents]
      summary: Delete agent
      description: Permanently delete an agent
      operationId: deleteAgent
      parameters:
        - $ref: '#/components/parameters/agentId'
      responses:
        '204':
          description: Agent deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /agents/{agentId}/pause:
    post:
      tags: [Agents]
      summary: Pause agent
      description: Temporarily pause agent operations
      operationId: pauseAgent
      parameters:
        - $ref: '#/components/parameters/agentId'
      responses:
        '200':
          description: Agent paused
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
  /agents/{agentId}/resume:
    post:
      tags: [Agents]
      summary: Resume agent
      description: Resume a paused agent
      operationId: resumeAgent
      parameters:
        - $ref: '#/components/parameters/agentId'
      responses:
        '200':
          description: Agent resumed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'

  # ==========================================================================
  # TRUST
  # ==========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    agentId:
      name: agentId
      in: path
      required: true
      schema:
        type: string
      description: Unique agent identifier
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      required: [success, error, requestId, timestamp]
      properties:
        success:
          type: boolean
          const: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
    TrustTier:
      type: integer
      minimum: 0
      maximum: 7
      description: |
        Trust tier (0-7):
        - 0: T0_SANDBOX (0-199)
        - 1: T1_OBSERVED (200-349)
        - 2: T2_PROVISIONAL (350-500)
        - 3: T3_MONITORED (501-649)
        - 4: T4_STANDARD (650-799)
        - 5: T5_TRUSTED (800-875)
        - 6: T6_CERTIFIED (876-950)
        - 7: T7_AUTONOMOUS (951-1000)
...
    Agent:
      type: object
      required: [id, name, description, ownerId, trustScore, trustTier, status, capabilities, executions, successRate, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        ownerId:
          type: string
        trustScore:
          type: number
          minimum: 0
          maximum: 1000
        trustTier:
          $ref: '#/components/schemas/TrustTier'
        tierName:
          type: string
        status:
          type: string
          enum: [ACTIVE, PAUSED, SUSPENDED, TERMINATED]
        capabilities:
          type: array
          items:
            type: string
        executions:
          type: integer
        successRate:
          type: number
          minimum: 0
          maximum: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        metadata:
          type: object
...
    UpdateAgentRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [ACTIVE, PAUSED]
        metadata:
          type: object
    AgentResponse:
      type: object
      required: [success, data, requestId, timestamp]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/Agent'
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
...